Add age_threshold to precache config and documentation
n.b., This includes quick-hack functionality to specify optional
configuration options for configparser to fetch. It would be better if
they were contingent options (i.e., we only need this, if that is set to
"foo", etc.), but it's not a huge deal and this works for us. The
documentation reflects this.